requirements.txt
pathway[llm]
google-generativeai
python-dotenv
tiktoken
litellm  # Crucial: Pathway uses this to talk to non-OpenAI models